Step2-3 요구사항 충족
- 댓글 기능 추가
- 댓글 삭제 기능 추가
- 게시글에 타인의 댓글이 달리면 게시글 삭제 방지 기능 추가
- 게시글에 자신의 댓글만 있으면 삭제 가능 기능 추가
- 게시글 삭제 시 delete플래그

+++ src/main/java/com/kakao/cafe/controller/ArticleController.java
@@ -2,8 +2,10 @@ package com.kakao.cafe.controller;
 
 import com.kakao.cafe.annotation.SessionCheck;
 import com.kakao.cafe.dto.ArticleDTO;
+import com.kakao.cafe.dto.ReplyDTO;
 import com.kakao.cafe.dto.UserDTO;
 import com.kakao.cafe.service.ArticleService;
+import com.kakao.cafe.service.ReplyService;
 import com.kakao.cafe.util.Constants;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Controller;
@@ -13,29 +15,33 @@ import org.springframework.web.bind.annotation.*;
 import javax.annotation.Resource;
 import javax.servlet.http.HttpSession;
 import javax.validation.Valid;
+import java.util.List;
 import java.util.Objects;
 
 @Slf4j
 @Controller
+@RequestMapping("/articles")
 public class ArticleController {
+    private static final String ARTICLE_UPDATE_NOT_ALLOWED_MESSAGE = "다른사용자가 작성한 내용은 수정할 수 없습니다.";
+    private static final String CANNOT_DELETE_ARTICLE_MESSAGE = "타인의 댓글이 달린 글은 삭제할 수 없습니다.";
     @Resource(name = "articleService")
     private ArticleService articleService;
-
-    private static final String ARTICLE_UPDATE_NOT_ALLOWED_MESSAGE = "다른사용자가 작성한 글은 수정할 수 없습니다.";
+    @Resource(name = "replyService")
+    private ReplyService replyService;
 
     @GetMapping
     String posts(Model model) {
         model.addAttribute("articles", articleService.getArticleList());
-        return "index";
+        return "article/list";
     }
 
-    @GetMapping("/articles/form")
+    @GetMapping("/form")
     @SessionCheck
     String form(HttpSession session) {
         return "article/form";
     }
 
-    @PostMapping("/articles")
+    @PostMapping
     @SessionCheck
     String articles(HttpSession session, @Valid ArticleDTO article, Model model) {
         if (articleService.insertArticle(article) < 1) {
@@ -47,9 +53,9 @@ public class ArticleController {
         return "redirect:/";
     }
 
-    @GetMapping("/articles/{id}/form")
+    @GetMapping("/{id}/form")
     @SessionCheck
-    String getArticleForm(@PathVariable long id, Model model, HttpSession session) {
+    String getArticleForm(HttpSession session, @PathVariable long id, Model model) {
         UserDTO user = (UserDTO) session.getAttribute("sessionUser");
         ArticleDTO article = articleService.getArticleById(id);
         if (!Objects.equals(article.getWriterId(), user.getId())) {
@@ -62,9 +68,9 @@ public class ArticleController {
         return "article/updateForm";
     }
 
-    @PutMapping("/articles/{id}/update")
+    @PutMapping("/{id}/update")
     @SessionCheck
-    String updateArticle(@PathVariable long id, @Valid ArticleDTO article, Model model, HttpSession session) {
+    String updateArticle(HttpSession session, @PathVariable long id, @Valid ArticleDTO article, Model model) {
         UserDTO sessionUser = (UserDTO) session.getAttribute("sessionUser");
         ArticleDTO articleInfo = articleService.getArticleById(id);
         if (!Objects.equals(articleInfo.getWriterId(), sessionUser.getId())) {
@@ -79,16 +85,24 @@ public class ArticleController {
         return "redirect:/";
     }
 
-    @DeleteMapping("/articles/{id}/delete")
+    @DeleteMapping("/{id}/delete")
     @SessionCheck
-    String deleteArticle(@PathVariable long id, Model model, HttpSession session) {
+    String deleteArticle(HttpSession session, @PathVariable long id, Model model) {
         UserDTO sessionUser = (UserDTO) session.getAttribute("sessionUser");
         ArticleDTO article = articleService.getArticleById(id);
         if (!Objects.equals(article.getWriterId(), sessionUser.getId())) {
             model.addAttribute(Constants.ERROR_MESSAGE_ATTRIBUTE_NAME, ARTICLE_UPDATE_NOT_ALLOWED_MESSAGE);
             return Constants.ERROR_PAGE_NAME;
         }
+        if (replyService.getOtherUserRepliesCount(id, sessionUser.getId()) > 0) {
+            model.addAttribute(Constants.ERROR_MESSAGE_ATTRIBUTE_NAME, CANNOT_DELETE_ARTICLE_MESSAGE);
+            return Constants.ERROR_PAGE_NAME;
+        }
         if (articleService.deleteArticle(id) <= 0) {
+            model.addAttribute(Constants.ERROR_MESSAGE_ATTRIBUTE_NAME, Constants.DEFAULT_ERROR_MESSAGE);
+            return Constants.ERROR_PAGE_NAME;
+        }
+        if (replyService.deleteAllReplies(id) < 0) {
             model.addAttribute(Constants.ERROR_MESSAGE_ATTRIBUTE_NAME, Constants.DEFAULT_ERROR_MESSAGE);
             return Constants.ERROR_PAGE_NAME;
         }
@@ -96,17 +110,54 @@ public class ArticleController {
         return "redirect:/";
     }
 
-    @GetMapping("/articles/{articleId}")
+    @GetMapping("/{articleId}")
     @SessionCheck
     String show(HttpSession session, @PathVariable long articleId, Model model) {
         UserDTO user = (UserDTO) session.getAttribute("sessionUser");
         articleService.increaseViews(articleId);
         ArticleDTO article = articleService.getArticleById(articleId);
+        List<ReplyDTO> replies = replyService.getArticleReplies(articleId, user.getId());
         model.addAttribute("isOwner", Objects.equals(article.getWriterId(), user.getId()));
         model.addAttribute("article", article);
         model.addAttribute("articleId", articleId);
+        model.addAttribute("replies", replies);
         log.info("get Article -> articleId : {}", articleId);
         return "article/show";
     }
+
+    @PostMapping("/{articleId}/reply/")
+    @SessionCheck
+    String insertReply(HttpSession session, @PathVariable long articleId, @Valid ReplyDTO reply, Model model) {
+        UserDTO user = (UserDTO) session.getAttribute("sessionUser");
+        long userId = user.getId();
+        reply.setWriterID(userId);
+        if (replyService.insertReply(reply) < 1) {
+            model.addAttribute(Constants.ERROR_MESSAGE_ATTRIBUTE_NAME, Constants.DEFAULT_ERROR_MESSAGE);
+            return Constants.ERROR_PAGE_NAME;
+        }
+        log.info("create Reply -> Article : {}, Writer : {}", articleId, userId);
+        return "redirect:/articles/" + articleId;
+    }
+
+    @DeleteMapping("/{articleId}/reply/{replyId}")
+    @SessionCheck
+    String deleteReply(HttpSession session, @PathVariable long articleId, @PathVariable long replyId, Model model) {
+
+        UserDTO sessionUser = (UserDTO) session.getAttribute("sessionUser");
+        long userId = sessionUser.getId();
+        ReplyDTO reply = replyService.getReplyById(userId, replyId);
+        if (!reply.getIsOwner()) {
+            model.addAttribute(Constants.ERROR_MESSAGE_ATTRIBUTE_NAME, ARTICLE_UPDATE_NOT_ALLOWED_MESSAGE);
+            return Constants.ERROR_PAGE_NAME;
+        }
+        if (replyService.deleteReply(replyId) <= 0) {
+            model.addAttribute(Constants.ERROR_MESSAGE_ATTRIBUTE_NAME, Constants.DEFAULT_ERROR_MESSAGE);
+            return Constants.ERROR_PAGE_NAME;
+        }
+        log.info("delete Reply -> ID : {}, Article : {}, Writer : {}", replyId, articleId, userId);
+
+        return "redirect:/articles/" + articleId;
+    }
+
 
 }
